+++ ecommerce/frontend/src/components/pages/SearchPage.less
@@ -1,18 +1,18 @@
 .search-page {
     display: flex;
-    gap: 20px;
+    flex-direction: column;
+    gap: 40px;
     margin: 20px;
+    margin-top: 40px;
 }
 
-.search-page-side-panel {
-    min-width: 30%;
-    height: fit-content;
-    position: sticky;
-    top: 0;
+.search-page-bar {
+    width: 50%;
+    margin: auto;
 }
 
-.product-grid {
+.search-page-main {
     display: flex;
-    flex-wrap: wrap;
-    gap: 10px;
+    gap: 20px;
 }
+
