+++ ecommerce/frontend/src/components/search/SearchResults.less
@@ -1,0 +1,6 @@
+
+.product-grid {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px;
+}
