+++ ecommerce/frontend/src/components/search/SearchSidePanel.less
@@ -1,0 +1,6 @@
+.search-page-side-panel {
+    min-width: 30%;
+    height: fit-content;
+    position: sticky;
+    top: 0;
+}
